Design: 넛지 카드 페이지 UI 리디자인
